% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.args.same.R
\name{assert_not_primitive}
\alias{assert_not_primitive}
\title{Checks if arguments are primitive functions}
\usage{
assert_not_primitive(...)
}
\arguments{
\item{...}{One or more functions}
}
\value{
logical - TRUE if all arguments are non-primitive functions
}
\description{
Checks if arguments are primitive functions
}
\details{
Wrapper for assertthat::assert_that()
}
