% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_args.R
\name{copy_args}
\alias{copy_args}
\title{Copies the arguments of one function to another}
\usage{
copy_args(from, to)
}
\arguments{
\item{from}{A function with the arguments to copy}

\item{to}{A function to receive new arguments}
}
\value{
function
}
\description{
Copies the arguments of  one function, "from", to "to".
}
\seealso{
See \code{\link{tictocify}} function body for a use case.
}
