% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_call_piped.R
\name{find_call_piped}
\alias{find_call_piped}
\title{Finds the call that generated the value that was passed through a pipe.}
\usage{
find_call_piped(.piped)
}
\arguments{
\item{.piped}{A value passed through a pipe.}
}
\value{
call
}
\description{
This function that receives a value through a pipe, then traces it through the stack to
find the original pipe call and returns it. The usage for this is when you want to write
a function that receives code rather than a value. You can put it as the first line of code
in your function to enable code to be piped. However, this defeats the purpose of the pipe
and can cause confusion for someone looking at your code.
}
