% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frite.R
\docType{package}
\name{frite}
\alias{frite}
\alias{frite-package}
\title{frite: A package for function modification and writing.}
\description{
This package contains functions that modify functions, helper functions for the
modifying functions, and diagnostic functions.
}
\section{Modifying functions}{

\code{\link{line_assign}}

\code{\link{line_insert}}

\code{\link{line_remove}}

\code{\link{copy_args}}
}

\section{Helper functions}{

\code{\link{plot_body}}

\code{\link{list_body}}
}

\section{Diagnostic functions}{

\code{\link{tictocify}}

\code{\link{is.output.same}}

\code{\link{is.args.same}}
}

