% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.args.same.R
\name{is.args.same}
\alias{is.args.same}
\title{Checks if two functions have identical arguments}
\usage{
is.args.same(.f1, .f2)
}
\arguments{
\item{.f1}{A function}

\item{.f2}{A function}
}
\value{
logical - TRUE if arguments are identical
}
\description{
Checks if two functions have identical arguments
}
