% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_remove.R
\name{line_remove}
\alias{line_remove}
\title{Removes an expression from a function body}
\usage{
line_remove(.f, line)
}
\arguments{
\item{.f}{A non-primitive function}

\item{line}{An expression from the body of the function}
}
\value{
function
}
\description{
This function will remove an expression from the function body. Use list_body() on any
function to see which lines are able to be removed.
}
\examples{

line_remove(.f = map, line = 2)
line_remove(.f = map, line = 3)

}
\seealso{
[list_body()]
}
