% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_body.R
\name{list_body}
\alias{list_body}
\title{Converts the body of a function to a list}
\usage{
list_body(.f)
}
\arguments{
\item{.f}{A non-primitive function}
}
\value{
A list
}
\description{
This allows for easy modification of functions and will help you see what code you are
changing when using other functions in this package.
}
