% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_insert.R
\name{line_insert}
\alias{line_insert}
\title{Inserts code into function}
\usage{
line_insert(.f, after_line, quoted_code)
}
\arguments{
\item{.f}{Any non-primitive function}

\item{after_line}{An integer that specifies where to insert quoted_code}

\item{quoted_code}{An expression wrapped in quote()}
}
\value{
function
}
\description{
Inserts code wrapped in the quote function into a function after the specified line.
}
\examples{

lapply_hello <- line_insert(lapply, 1, quote(print('Hello!')))

}
\seealso{
\code{\link{list_body}} \code{\link{line_insert}} \code{\link{line_remove}}
}
