% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_body.R
\name{plot_body}
\alias{plot_body}
\title{Plots the body of a non-primitive function}
\usage{
plot_body(.f, spacing = 0.15, font_size = 0.8, line_width = 80)
}
\arguments{
\item{.f}{A function}

\item{spacing}{Specifies space between lines, positive number}

\item{font_size}{Specifies text size, positive number}

\item{line_width}{Specifies the line width, positive number}
}
\value{
text in plot window and function
}
\description{
Plots the body of a function so that you can easily see it while using other functions
in this package. It is pipeable and will return the function that you put into it.
}
\examples{

plot_body(map)
plot_body(cross, spacing = .0872, font_size = .54, line_width = 100)

}
\seealso{
[list_body()]
}
