% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictocify.R
\name{tictocify}
\alias{tictocify}
\title{Creates a function that functions identically to the original, but gives execution time}
\usage{
tictocify(..f, tic_args = NULL, toc_args = NULL)
}
\arguments{
\item{..f}{Any function that isn't primitive (sum, list, etc...)}

\item{tic_args}{A list of arguments to be passed into tic()}

\item{toc_args}{A list of arguments to be passed into toc()}
}
\value{
function
}
\description{
Takes any function and returns a nearly identical function that will behave in the same
manner, but will return the time it takes to execute the function.
}
\details{
Creates a wrapper function around a function call, preceded by tic() and
 toc(). The wrapper will be given the same arguments as the original function and
 work identically to the original function.
}
\examples{

set.seed(1)
stuff <- rnorm(n = 100000)

map_lgl_new <- tictocify(map_lgl)

map_lgl_new(.f = function(x) x > 0, .x = stuff)
map_lgl_new(stuff, function(x) x > 0) \%>\% sum() \%>\% as.character()

is.output.same(map_lgl_new(stuff, function(x) x > 0), map_lgl)

}
\seealso{
[is.output.same()]
}
